Split manage workspace orchestration modules
