+++ src/features/workbench/manage-workspace.tsx
@@ -1,8 +1,6 @@
 import type { ReactNode } from "react";
 
-import PortfolioScreenRail, {
-  type PortfolioScreenRailModeItem,
-} from "@/apps/portfolio/components/portfolio-screen-rail";
+import PortfolioScreenRail from "@/apps/portfolio/components/portfolio-screen-rail";
 import {
   AppPageShell,
   DefinitionList,
@@ -42,253 +40,16 @@ import {
   toneForState,
 } from "@/features/workbench/manage-workspace-view-model";
 import {
-  getDpmCommandCenter,
-  getDpmCommandCenterExceptions,
-  getDpmMandateByPortfolio,
-  getDpmMandateHealth,
-  getDpmOutcomeReviews,
-  getDpmPmOperatingQualityFairnessAnalysis,
-  getDpmPortfolioMemory,
-  getDpmProofPack,
-  listDpmCampaignDiscovery,
-  listDpmCampaignDefinitions,
-  listDpmPmOperatingQualityFairnessAnalyses,
-  getPortfolio360,
-  listDpmPmOperatingQualityPolicies,
-  listDpmPmOperatingQualityScoreRuns,
-  listDpmWaves,
-} from "@/features/workbench/api";
+  buildManageModeHref,
+  buildManageModeItems,
+  getManageModeDefinition,
+  type ManageMode,
+} from "@/features/workbench/manage-workspace-navigation";
+import {
+  readDpmMandateId,
+  type ManageWorkspaceData,
+} from "@/features/workbench/manage-workspace-data";
 import styles from "./manage-workspace.module.css";
-
-export type ManageMode =
-  | "overview"
-  | "mandate"
-  | "waves"
-  | "construction"
-  | "memory"
-  | "quality"
-  | "reviews"
-  | "proof";
-
-export type ManageWorkspaceData = {
-  portfolio: Awaited<ReturnType<typeof getPortfolio360>>;
-  commandCenter: Awaited<ReturnType<typeof getDpmCommandCenter>> | null;
-  commandCenterExceptions: Awaited<ReturnType<typeof getDpmCommandCenterExceptions>> | null;
-  mandate: Awaited<ReturnType<typeof getDpmMandateByPortfolio>> | null;
-  mandateHealth: Awaited<ReturnType<typeof getDpmMandateHealth>> | null;
-  commandCenterError: string | null;
-  portfolioMemory: Awaited<ReturnType<typeof getDpmPortfolioMemory>> | null;
-  portfolioMemoryError: string | null;
-  pmOperatingQualityPolicies: Awaited<ReturnType<typeof listDpmPmOperatingQualityPolicies>> | null;
-  pmOperatingQualityPoliciesError: string | null;
-  pmOperatingQualityScoreRuns: Awaited<ReturnType<typeof listDpmPmOperatingQualityScoreRuns>> | null;
-  pmOperatingQualityScoreRunsError: string | null;
-  pmOperatingQualityFairnessAnalyses: Awaited<
-    ReturnType<typeof listDpmPmOperatingQualityFairnessAnalyses>
-  > | null;
-  pmOperatingQualityFairnessAnalysesError: string | null;
-  pmOperatingQualityFairnessAnalysisDetail: Awaited<
-    ReturnType<typeof getDpmPmOperatingQualityFairnessAnalysis>
-  > | null;
-  pmOperatingQualityFairnessAnalysisDetailError: string | null;
-  waves: Awaited<ReturnType<typeof listDpmWaves>> | null;
-  wavesError: string | null;
-  campaignDefinitions: Awaited<ReturnType<typeof listDpmCampaignDefinitions>> | null;
-  campaignDefinitionsError: string | null;
-  campaignDiscovery: Awaited<ReturnType<typeof listDpmCampaignDiscovery>> | null;
-  campaignDiscoveryError: string | null;
-  outcomeReviews: Awaited<ReturnType<typeof getDpmOutcomeReviews>> | null;
-  outcomeReviewError: string | null;
-  proofPack: Awaited<ReturnType<typeof getDpmProofPack>> | null;
-  proofPackError: string | null;
-};
-
-export const MANAGE_MODE_DEFINITIONS: Array<{
-  key: ManageMode;
-  label: string;
-  detail: string;
-  title: string;
-  description: string;
-}> = [
-  {
-    key: "overview",
-    label: "Overview",
-    detail: "Mandate posture",
-    title: "Manage Overview",
-    description: "Discretionary mandate readiness and rebalance posture.",
-  },
-  {
-    key: "mandate",
-    label: "Mandate",
-    detail: "Health and exceptions",
-    title: "Mandate Health",
-    description: "Mandate readiness, attention items, and recommended actions.",
-  },
-  {
-    key: "waves",
-    label: "Rebalance",
-    detail: "Rebalance lifecycle",
-    title: "Rebalance Waves",
-    description: "Rebalance simulation, approval readiness, and execution handoff.",
-  },
-  {
-    key: "construction",
-    label: "Construction",
-    detail: "Alternatives",
-    title: "Construction Alternatives",
-    description: "Supported construction alternatives for advisor and PM review.",
-  },
-  {
-    key: "memory",
-    label: "Memory",
-    detail: "Portfolio memory",
-    title: "Portfolio Memory",
-    description: "Portfolio decision memory and operating history.",
-  },
-  {
-    key: "quality",
-    label: "PM Quality",
-    detail: "Operating quality",
-    title: "PM Operating Quality",
-    description: "Governance review of Manage-owned PM quality policy and evidence.",
-  },
-  {
-    key: "reviews",
-    label: "Reviews",
-    detail: "Outcome review",
-    title: "Outcome Reviews",
-    description: "Post-trade review and realized-versus-expected variance.",
-  },
-  {
-    key: "proof",
-    label: "Evidence",
-    detail: "Decision evidence",
-    title: "Evidence Pack",
-    description: "Mandate evidence, approval readiness, and client handoff support.",
-  },
-];
-
-export async function loadManageWorkspaceData(
-  portfolio: Awaited<ReturnType<typeof getPortfolio360>>
-): Promise<ManageWorkspaceData> {
-  const portfolioId = portfolio.portfolio.portfolio_id;
-  const [
-    commandCenterResult,
-    exceptionsResult,
-    mandateResult,
-    memoryResult,
-    wavesResult,
-    campaignDefinitionsResult,
-    campaignDiscoveryResult,
-    pmQualityPoliciesResult,
-    pmQualityScoreRunsResult,
-    pmQualityFairnessAnalysesResult,
-    reviewsResult,
-  ] = await Promise.allSettled([
-    getDpmCommandCenter({ limit: 25 }),
-    getDpmCommandCenterExceptions({ state: "ACTIVE", limit: 25 }),
-    getDpmMandateByPortfolio(portfolioId),
-    getDpmPortfolioMemory({ portfolioId, limit: 100 }),
-    listDpmWaves({ triggerType: "EXPLICIT_PORTFOLIO_LIST", limit: 10 }),
-    listDpmCampaignDefinitions({ campaignStatus: "ACTIVE", limit: 10 }),
-    listDpmCampaignDiscovery({ campaignStatus: "ACTIVE", limit: 10 }),
-    listDpmPmOperatingQualityPolicies({ limit: 10 }),
-    listDpmPmOperatingQualityScoreRuns({ limit: 10 }),
-    listDpmPmOperatingQualityFairnessAnalyses({ limit: 10 }),
-    getDpmOutcomeReviews({ portfolioId, limit: 10 }),
-  ]);
-
-  const mandate = readSettledValue(mandateResult);
-  const mandateId = readDpmMandateId(mandate?.data ?? null);
-  const outcomeReviews = readSettledValue(reviewsResult);
-  let mandateHealth: Awaited<ReturnType<typeof getDpmMandateHealth>> | null = null;
-  if (mandateId) {
-    try {
-      mandateHealth = await getDpmMandateHealth(mandateId);
-    } catch {
-      mandateHealth = null;
-    }
-  }
-  let proofPack: Awaited<ReturnType<typeof getDpmProofPack>> | null = null;
-  let proofPackError: string | null = null;
-  const proofPackId = readDpmProofPackId(outcomeReviews?.data ?? null);
-  if (proofPackId) {
-    try {
-      proofPack = await getDpmProofPack(proofPackId);
-    } catch (error) {
-      proofPackError = error instanceof Error ? error.message : "Evidence pack endpoint unavailable.";
-    }
-  }
-  const fairnessAnalyses = readSettledValue(pmQualityFairnessAnalysesResult);
-  let fairnessAnalysisDetail: Awaited<
-    ReturnType<typeof getDpmPmOperatingQualityFairnessAnalysis>
-  > | null = null;
-  let fairnessAnalysisDetailError: string | null = null;
-  const fairnessAnalysisId = readDpmFairnessAnalysisId(fairnessAnalyses?.data ?? null);
-  if (fairnessAnalysisId) {
-    try {
-      fairnessAnalysisDetail = await getDpmPmOperatingQualityFairnessAnalysis(fairnessAnalysisId);
-    } catch (error) {
-      fairnessAnalysisDetailError =
-        error instanceof Error
-          ? error.message
-          : "PM operating quality fairness-analysis detail endpoint unavailable.";
-    }
-  }
-
-  return {
-    portfolio,
-    commandCenter: readSettledValue(commandCenterResult),
-    commandCenterExceptions: readSettledValue(exceptionsResult),
-    mandate,
-    mandateHealth,
-    commandCenterError: readSettledError(
-      commandCenterResult,
-      "Mandate readiness is temporarily unavailable."
-    ),
-    portfolioMemory: readSettledValue(memoryResult),
-    portfolioMemoryError: readSettledError(
-      memoryResult,
-      "Portfolio-memory endpoint unavailable."
-    ),
-    pmOperatingQualityPolicies: readSettledValue(pmQualityPoliciesResult),
-    pmOperatingQualityPoliciesError: readSettledError(
-      pmQualityPoliciesResult,
-      "PM operating quality policy endpoint unavailable."
-    ),
-    pmOperatingQualityScoreRuns: readSettledValue(pmQualityScoreRunsResult),
-    pmOperatingQualityScoreRunsError: readSettledError(
-      pmQualityScoreRunsResult,
-      "PM operating quality score-run endpoint unavailable."
-    ),
-    pmOperatingQualityFairnessAnalyses: fairnessAnalyses,
-    pmOperatingQualityFairnessAnalysesError: readSettledError(
-      pmQualityFairnessAnalysesResult,
-      "PM operating quality fairness-analysis list endpoint unavailable."
-    ),
-    pmOperatingQualityFairnessAnalysisDetail: fairnessAnalysisDetail,
-    pmOperatingQualityFairnessAnalysisDetailError: fairnessAnalysisDetailError,
-    waves: readSettledValue(wavesResult),
-    wavesError: readSettledError(wavesResult, "DPM wave endpoint unavailable."),
-    campaignDefinitions: readSettledValue(campaignDefinitionsResult),
-    campaignDefinitionsError: readSettledError(
-      campaignDefinitionsResult,
-      "DPM campaign-definition endpoint unavailable."
-    ),
-    campaignDiscovery: readSettledValue(campaignDiscoveryResult),
-    campaignDiscoveryError: readSettledError(
-      campaignDiscoveryResult,
-      "DPM campaign-discovery endpoint unavailable."
-    ),
-    outcomeReviews,
-    outcomeReviewError: readSettledError(
-      reviewsResult,
-      "Outcome review endpoint unavailable."
-    ),
-    proofPack,
-    proofPackError,
-  };
-}
 
 export function ManageWorkspace({
   data,
@@ -934,44 +695,6 @@ function buildManageActivityRows(
       ];
 }
 
-function buildManageModeItems(portfolioId: string, activeMode: ManageMode): PortfolioScreenRailModeItem[] {
-  return MANAGE_MODE_DEFINITIONS.map((mode) => ({
-    key: mode.key,
-    label: mode.label,
-    detail: mode.detail,
-    active: activeMode === mode.key,
-    href: buildManageModeHref(portfolioId, mode.key),
-    title: mode.description,
-  }));
-}
-
-function buildManageModeHref(portfolioId: string, mode: ManageMode) {
-  const encoded = encodeURIComponent(portfolioId);
-  return mode === "overview" ? `/workbench/${encoded}` : `/workbench/${encoded}?mode=${mode}`;
-}
-
-export function normalizeManageMode(value: string | undefined): ManageMode {
-  const requested = value?.trim().toLowerCase();
-  return MANAGE_MODE_DEFINITIONS.some((mode) => mode.key === requested)
-    ? (requested as ManageMode)
-    : "overview";
-}
-
-function getManageModeDefinition(mode: ManageMode) {
-  return MANAGE_MODE_DEFINITIONS.find((definition) => definition.key === mode) ?? MANAGE_MODE_DEFINITIONS[0];
-}
-
-function readSettledValue<T>(result: PromiseSettledResult<T>): T | null {
-  return result.status === "fulfilled" ? result.value : null;
-}
-
-function readSettledError<T>(result: PromiseSettledResult<T>, fallback: string): string | null {
-  if (result.status === "fulfilled") {
-    return null;
-  }
-  return result.reason instanceof Error ? result.reason.message : fallback;
-}
-
 function formatAmount(value: number | null | undefined): string {
   if (value === null || value === undefined) {
     return "N/A";
@@ -998,77 +721,4 @@ function parseCount(value: number | string | null | undefined): number {
   }
   const parsed = Number.parseInt(value, 10);
   return Number.isFinite(parsed) ? parsed : 0;
-}
-
-export function readDpmMandateId(data: Record<string, unknown> | null): string | null {
-  if (!data) {
-    return null;
-  }
-  if (typeof data.mandate_id === "string" && data.mandate_id.trim().length > 0) {
-    return data.mandate_id;
-  }
-  const mandate = data.mandate;
-  if (mandate && typeof mandate === "object" && !Array.isArray(mandate)) {
-    const mandateId = (mandate as Record<string, unknown>).mandate_id;
-    return typeof mandateId === "string" && mandateId.trim().length > 0 ? mandateId : null;
-  }
-  return null;
-}
-
-export function readDpmProofPackId(data: Record<string, unknown> | null): string | null {
-  if (!data) {
-    return null;
-  }
-  if (typeof data.proof_pack_id === "string" && data.proof_pack_id.trim().length > 0) {
-    return data.proof_pack_id;
-  }
-  const items = Array.isArray(data.items) ? data.items : [];
-  for (const item of items) {
-    if (!item || typeof item !== "object" || Array.isArray(item)) {
-      continue;
-    }
-    const proofPackId = (item as Record<string, unknown>).proof_pack_id;
-    if (typeof proofPackId === "string" && proofPackId.trim().length > 0) {
-      return proofPackId;
-    }
-  }
-  return null;
-}
-
-export function readDpmFairnessAnalysisId(data: Record<string, unknown> | null): string | null {
-  if (!data) {
-    return null;
-  }
-  if (
-    typeof data.fairness_analysis_id === "string" &&
-    data.fairness_analysis_id.trim().length > 0
-  ) {
-    return data.fairness_analysis_id;
-  }
-  const fairnessAnalysis = data.fairness_analysis;
-  if (
-    fairnessAnalysis &&
-    typeof fairnessAnalysis === "object" &&
-    !Array.isArray(fairnessAnalysis)
-  ) {
-    const fairnessAnalysisId = (fairnessAnalysis as Record<string, unknown>).fairness_analysis_id;
-    if (typeof fairnessAnalysisId === "string" && fairnessAnalysisId.trim().length > 0) {
-      return fairnessAnalysisId;
-    }
-  }
-  const items = Array.isArray(data.fairness_analyses)
-    ? data.fairness_analyses
-    : Array.isArray(data.items)
-      ? data.items
-      : [];
-  for (const item of items) {
-    if (!item || typeof item !== "object" || Array.isArray(item)) {
-      continue;
-    }
-    const fairnessAnalysisId = (item as Record<string, unknown>).fairness_analysis_id;
-    if (typeof fairnessAnalysisId === "string" && fairnessAnalysisId.trim().length > 0) {
-      return fairnessAnalysisId;
-    }
-  }
-  return null;
 }
